requirements.txt
tensorflow==2.15.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
flask==3.0.0
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0
requests==2.31.0
tqdm==4.66.1
